+++ ATV. 08 - PAPEIS E POLITICAS (POLICES) DE USUÁRIO/app/Http/Controllers/PublisherController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Support\Facades\Gate;
 use App\Models\Publisher;
 use Illuminate\Http\Request;
 
@@ -10,17 +10,20 @@ class PublisherController extends Controller
 
     public function index()
     {
+        Gate::authorize('manage-publishers');
         $publishers = Publisher::all();
         return view('publishers.index', compact('publishers'));
     }
 
     public function create()
     {
+        Gate::authorize('manage-publishers');
         return view('publishers.create');
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('manage-publishers');
         $request->validate([
             'name' => 'required|string|unique:publishers|max:255',
             'address' =>'required|string|unique:publishers|max:255'
@@ -33,16 +36,19 @@ class PublisherController extends Controller
 
     public function show(Publisher $publisher)
     {
+        Gate::authorize('manage-publishers');
         return view('publishers.show', compact('publisher'));
     }
 
     public function edit(Publisher $publisher)
     {
+        Gate::authorize('manage-publishers');
         return view('publishers.edit', compact('publisher'));
     }
 
     public function update(Request $request, Publisher $publisher)
     {
+        Gate::authorize('manage-publishers');
         $request->validate([
             'name' => 'required|string|max:255',
             'address' =>'required|string|max:255'
@@ -55,6 +61,7 @@ class PublisherController extends Controller
 
     public function destroy(Publisher $publisher)
     {
+        Gate::authorize('manage-publishers');
         $publisher->delete();
 
         return redirect()->route('publishers.index')->with('sucess', 'Editora deletada com sucesso.');
